Throw exception in CDNClient if http request is not successful
Fixes #650
